migrated snippet-db module to ts. waiting on newest release of mysql2 for promise pool types to be patched
